changing data type for index to compare to size

+++ hash_tables/4-hash_table_get.c
@@ -10,7 +10,7 @@
 char *hash_table_get(const hash_table_t *ht, const char *key)
 {
 	hash_node_t *current;
-	int index;
+	long unsigned int index;
 
 	if (ht == NULL || key == NULL)
 		return (NULL);
